+++ pkg/dockerscript/scanner/extra.go
