+++ pkg/dockerscript/scanner/scanner.go
